agent/requirements.txt
langchain-groq
langgraph
langgraph-checkpoint-postgres
copilotkit
tavily-python
psycopg2-binary
psycopg[binary]
python-dotenv
